% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelViz.R
\name{mandelViz}
\alias{mandelViz}
\title{Title}
\usage{
mandelViz(
  real_mar = c(-1.6, 1),
  imag_mar = c(-1, 1),
  resol = 250,
  n_cores = parallel::detectCores() - 1,
  parallel = "yes",
  cmd = FALSE
)
}
\arguments{
\item{real_mar}{a vector of two numbers indicating the lower and upper limit
on the real axis.}

\item{imag_mar}{a vector of two numbers indicating the lower and upper limit
on the imaginary axis.}

\item{resol}{the number of equally spaced points on the smaller of the two
axes.}

\item{n_cores}{number of cores if the parallel function is chosen.}

\item{parallel}{default is "yes". It "no" a sapply is used for calculation}

\item{cmd}{logical: if the function is used in an executable file,
see details.}
}
\value{
a .tiff image stored in the current working directory, where the
function was used.
}
\description{
Title
}
